feature: enable user to delete their own posts

+++ spectrum/views.py
@@ -136,6 +136,9 @@ def updateInterest(request, pk):
 def deleteInterest(request, pk):
     interest = Interest.objects.get(id=pk)
 
+    if request.user != interest.host:
+        return HttpResponse('You are not allowed to perform this action')
+
     if request.method == 'POST':
         interest.delete()
         return redirect('home')
@@ -166,5 +169,17 @@ def quitInterest(request,pk):
         'interests':interests
     }
     return render(request, 'spectrum/home.html', context)
+
+
+def deletePost(request, pk):
+    post = Post.objects.get(id=pk)
+
+    if request.user != post.user:
+        return HttpResponse('You are not allowed to perform this action')
+
+    if request.method == 'POST':
+        post.delete()
+        return redirect('home')
+    return render(request, 'spectrum/delete.html', {'obj':post})
